Fixed resultant mess
